[fix] security on the server side

+++ SocialNetwork/Controllers/AuthorizationController.cs
@@ -37,6 +37,12 @@ namespace SocialNetwork.Controllers
         {
             //добавить валидацию id             
             var authorization = await unitOfWork.AuthorizationRepository.GetById(id);
+
+            if (User.Identity.Name != authorization.CredentialRef.ToString())
+            {
+                return Unauthorized();
+            }
+
             if (authorization != null)
             {
                 return Ok(authorization);
@@ -80,6 +86,10 @@ namespace SocialNetwork.Controllers
             var authorization = await unitOfWork.AuthorizationRepository.GetById(id);
             if (authorization != null)
             {
+                if (User.Identity.Name != authorization.CredentialRef.ToString())
+                {
+                    return Unauthorized();
+                }
                 await unitOfWork.AuthorizationRepository.Delete(authorization);
                 return Ok();
             }
@@ -93,6 +103,11 @@ namespace SocialNetwork.Controllers
         [Produces("application/json")]
         public async Task<ActionResult<ICollection<Authorization>>> GetAllAuthorizationByCredential([FromQuery]int credentialId)
         {
+            if (User.Identity.Name != credentialId.ToString())
+            {
+                return Unauthorized();
+            }
+
             if (credentialId != 0)
             {
                 var authorizations = await unitOfWork.AuthorizationRepository.GetAllAuthorizantionsByCredentialId(credentialId);
